==> scripts/1_simulation_analysis/6_collect_metrics.R <==
library(tidyverse)
source("scripts/functions.R")

data_dir <- "results/1_simulation_analysis"
results_dir <- "results/1_simulation_analysis/metrics"


# Collect data on the fraction of cells with non-zero expression of a gene and mean normalized expression of a gene ----
## Get list of files with normalized counts matrices
raw_counts_files <- list.files(file.path(data_dir, "gene_expression_filtered"), pattern = "norm.txt.gz$", full.names = TRUE)
imputed_counts_files <- list.files(file.path(data_dir, "gene_expression_imputed"), pattern = ".txt.gz$", full.names = TRUE)

## Calculate and aggregate metrics tables
metrics <- collect_metrics(c(raw_counts_files, imputed_counts_files), replace_na = FALSE)
dataset_size <- metrics$dataset_size
dropout_rate <- metrics$dropout_rate
nonzero_fraction <- metrics$nonzero_fraction
avg_norm_counts <- metrics$avg_norm_counts


## Save metrics into files -----------------------------------------------------
write_tsv(dataset_size, file.path(results_dir, "dataset_size.txt"))
write_tsv(dropout_rate, file.path(results_dir, "dropout_rate.txt"))
write_tsv(nonzero_fraction, file.path(results_dir, "nonzero_fraction.txt"))
write_tsv(avg_norm_counts, file.path(results_dir, "avg_norm_counts.txt"))


# Calculate NRMSE values to access the recovery of gene expression levels ------
calculate_nrmse <- function(true_mtx_file, observed_mtx_file) {
  # Extract dataset name (simulation method and cell type in the reference data)
  sample_name = str_extract(basename(true_mtx_file), "[a-z-]+_[a-z]+")
  
  # Extract info on imputation method (change to 'observed' if it is data generated by simulation)
  method = basename(observed_mtx_file) %>% 
    # Remove file format
    str_remove(".txt.gz") %>%
    # Remove sample ID
    str_remove_all(paste0(str_split_i(basename(observed_mtx_file), pattern = "\\.", i = 1), "."))
  if (method == "norm") { method = "observed" }
  
  # Read in expression matrices from files
  observed_mtx = read.table(observed_mtx_file, header = TRUE, row.names = 1)
  true_mtx = read.table(true_mtx_file, header = TRUE, row.names = 1)
  
  # Subset matrices to contain only shared set of cells and genes
  shared_cells = intersect(colnames(observed_mtx), colnames(true_mtx))
  shared_genes = intersect(rownames(observed_mtx), rownames(true_mtx))
  true_mtx = true_mtx[shared_genes, shared_cells]
  observed_mtx = observed_mtx[shared_genes, shared_cells]
  
  # RMSE for each gene (in rows)
  rmse = sqrt(rowMeans((true_mtx - observed_mtx)^2))
  
  # Normalize RMSE by mean true expression
  nrmse = as.data.frame(rmse / rowMeans(true_mtx))
  
  # Add dataset name and imputation method as a column name
  colnames(nrmse) = paste(sample_name, method, sep = ".")
  
  # Add a column with gene names
  nrmse = rownames_to_column(nrmse, "GeneID")
  
  return(nrmse)
}


## Create data frame with files of true and observed/imputed data -------------- 
nrmse_scheme <- tibble(
  true_mtx_file = str_subset(raw_counts_files, "true_counts"),
  sample_name = str_extract(basename(true_mtx_file), "[a-z-]+_[a-z]+")) %>% 
  rowwise() %>% 
  mutate(obs_or_imputed_mtx_file = list(str_subset(c(str_subset(raw_counts_files, "obs_counts"), imputed_counts_files), sample_name))) %>% 
  unnest(obs_or_imputed_mtx_file)


## Create data frame with NRMSE values for all datasets ------------------------
nrmse_df <- map2(nrmse_scheme$true_mtx_file, nrmse_scheme$obs_or_imputed_mtx_file, calculate_nrmse) %>% 
  reduce(full_join, by = "GeneID")

write_tsv(nrmse_df, file.path(results_dir, "nrmse.txt"))
